pyproject.toml
[project]
name = "live-trading-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.13.2",
    "eth-account>=0.13.7",
    "fastapi>=0.120.2",
    "hyperliquid-python-sdk>=0.20.0",
    "jinja2>=3.1.6",
    "numpy>=2.0.2",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.6.1",
    "ta>=0.11.0",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.38.0",
]
